+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * --include-container_roots in service products-contain-component
+* --exclude-unreleased in service products-contain-component
 
 ## [0.4.2] - 2024-01-14
 ### Added